moving config
